完成json的读入部分 完成大多数主逻辑判定，添加yaml-cpp支持 添加yaml配置文件 添加QT GUI的支持 添加QT GUI配置工具窗体

+++ codes/main/configs/NukeConfiguration.cpp
@@ -27,12 +27,12 @@ double NukeConfiguration::getRadius() const{
     return radius;
 }
 
-int NukeConfiguration::getSpeed() const{
+double NukeConfiguration::getSpeed() const{
     return speed;
 }
 
 NukeConfiguration::NukeConfiguration(const int aoeDamage, const double aoeDamageRadius,
- const int damageDecrease, const double radius, const int speed) {
+ const int damageDecrease, const double radius, const double speed) {
     this->aoeDamage = aoeDamage;
     this->aoeDamageRadius = aoeDamageRadius;
     this-> damageDecrease = damageDecrease;
