+++ codes/main/configs/WeaponConfiguration.cpp
@@ -19,7 +19,7 @@ int WeaponConfiguration::getDamage() const{
     return damage;
 }
 
-int WeaponConfiguration::getSpeed() const{
+double WeaponConfiguration::getSpeed() const{
     return speed;
 }
 
@@ -27,7 +27,7 @@ int WeaponConfiguration::getDelay() const{
     return delay;
 }
 
-WeaponConfiguration::WeaponConfiguration(const double radius, const int damage, const int speed, const int delay) {
+WeaponConfiguration::WeaponConfiguration(const double radius, const int damage, const double speed, const int delay) {
     this->radius = radius;
     this->damage =damage;
     this->speed =speed;
